% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customControls.R
\name{goButton}
\alias{goButton}
\title{Coloured action button
https://gist.github.com/xiaodaigh/7012930}
\usage{
goButton(inputId, label, btn.style = "", css.class = "")
}
\description{
Coloured action button
https://gist.github.com/xiaodaigh/7012930
}

